refactor(chat): extract chat-history builders + token counting to a part file
Move the private prompt chat-history formatters and token accounting
(`_buildChatHistory`, `_buildChatHistoryWithBudget`, `_countTokens`) into
chat_service_history.dart as a `part of` extension. Pure formatting over
`_messages` + token budgeting, no engine/orchestration logic; verbatim move.

Note: the static `_macroPattern` had to be qualified as
`ChatService._macroPattern` inside the extension — static members of the
extended type cannot be referenced unqualified from an extension.

chat_service.dart 9,931 → 9,852 lines (10,506 at the start of decomposition).
flutter analyze clean; unit suite (1334) green.
